requirements.txt
requests
fastapi
uvicorn
python-multipart

bcrypt==3.2.2
passlib[bcrypt]
python-jose[cryptography]

sqlalchemy
aiosqlite
PyJWT